// fem-vision-studio/design-mockup/src/lib/types/plaat/PlaatCombinatie.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PlaatElementSpanning } from "./PlaatElementSpanning";

/**
 * De elementspanningen van één UGT-combinatie.
 */
export type PlaatCombinatie = { combination_id: number, elements: Array<PlaatElementSpanning>, };